--- hopsworks-common/src/main/java/io/hops/hopsworks/common/user/ldap/LdapGroupMapper.java
@@ -1,121 +1,0 @@
-/*
- * Changes to this file committed after and not including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
- * are released under the following license:
- *
- * This file is part of Hopsworks
- * Copyright (C) 2018, Logical Clocks AB. All rights reserved
- *
- * Hopsworks is free software: you can redistribute it and/or modify it under the terms of
- * the GNU Affero General Public License as published by the Free Software Foundation,
- * either version 3 of the License, or (at your option) any later version.
- *
- * Hopsworks is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- * PURPOSE.  See the GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License along with this program.
- * If not, see <https://www.gnu.org/licenses/>.
- *
- * Changes to this file committed before and including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
- * are released under the following license:
- *
- * Copyright (C) 2013 - 2018, Logical Clocks AB and RISE SICS AB. All rights reserved
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify, merge,
- * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
- * persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS  OR IMPLIED, INCLUDING
- * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL  THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR  OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package io.hops.hopsworks.common.user.ldap;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.StringTokenizer;
-
-public class LdapGroupMapper {
-
-  private static final String GROUP_SEPARATOR = ",";
-  private static final String MAPPING_SEPARATOR = "->";
-  private static final String GROUP_MAPPING_SEPARATOR = ";";
-  private final String mappingStr;
-  private final Map<String, List<String>> mappings = new HashMap<>();
-
-  public LdapGroupMapper(String mappings) {
-    this.mappingStr = mappings;
-    parse(mappings);
-  }
-
-  /**
-   * Mapping string in format 
-   * "groupA-> HOPS_USER,HOPS_ADMIN;group B -> HOPS_USER; group-C-> HOPS_ADMIN,CLUSTER_AGENT, AGENT"
-   * @param mappingStr 
-   */
-  private void parse(String mappingStr) {
-    if (mappingStr == null || mappingStr.isEmpty()) {
-      return;
-    }
-    StringTokenizer tokenizer = new StringTokenizer(mappingStr, GROUP_MAPPING_SEPARATOR);
-    while (tokenizer.hasMoreElements()) {
-      String mapping = tokenizer.nextToken();
-      String[] mappingGroups = mapping.split(MAPPING_SEPARATOR);
-      String mappedGroup = null;
-      String[] mappedToGroups = null;
-      if (mappingGroups != null && mappingGroups.length == 2) {
-        mappedGroup = mappingGroups[0].trim();
-        mappedToGroups = mappingGroups[1].split(GROUP_SEPARATOR);
-      }
-      if (mappedGroup == null || mappedGroup.isEmpty() || mappedToGroups == null || mappedToGroups.length < 1) {
-        continue;
-      }
-      List<String> mappedTOGroupList = new ArrayList<>();
-      for (String grp : mappedToGroups) {
-        mappedTOGroupList.add(grp.trim());
-      }
-      this.mappings.put(mappedGroup, mappedTOGroupList);
-    }
-  }
-
-  public Map<String, List<String>> getMappings() {
-    return mappings;
-  }
-
-  public List<String> getMappedGroups(List<String> groups) {
-    List<String> mappedGroups = new ArrayList<>();
-    if (this.mappings == null || this.mappings.isEmpty() || groups == null || groups.isEmpty()) {
-      return mappedGroups;
-    }
-    for (String group : groups) {
-      addUnique(this.mappings.get(group), mappedGroups);
-    }
-    return mappedGroups;
-  }
-
-  public String getMappingStr() {
-    return mappingStr;
-  }
-
-  private void addUnique(List<String> src, List<String> dest) {
-    if (src == null) {
-      return;
-    }
-    for (String str : src) {
-      if (!dest.contains(str)) {
-        dest.add(str);
-      }
-    }
-  }
-
-}
